Move middleware to seperate file
